+++ app/controllers/users/registrations_controller.rb
@@ -2,6 +2,7 @@ module Users
   class RegistrationsController < Devise::RegistrationsController
     skip_before_action :set_current_user
     skip_before_action :save_current_path
+    skip_before_action :email_confirmed?
     before_action :save_omniauth_login_locale, only: :new
 
     private