+++ app/mailers/user_mailer.rb
@@ -1,0 +1,8 @@
+# Emails for users
+class UserMailer < ApplicationMailer
+  def confirmation_email(user_id:)
+    @user = User.find_by(id: user_id)
+    return if @user.nil?
+    mail(to: @user.email, subject: t('mailer.user.confirmation_email.subject'))
+  end
+end