+++ db/migrate/20191120220009_update_users_confirmations.rb
@@ -1,0 +1,5 @@
+class UpdateUsersConfirmations < ActiveRecord::Migration[5.2]
+  def change
+    User.update_all(confirmed_at: DateTime.now)
+  end
+end