+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_11_17_204452) do
+ActiveRecord::Schema.define(version: 2019_11_20_220009) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_trgm"
@@ -291,6 +291,8 @@ ActiveRecord::Schema.define(version: 2019_11_17_204452) do
     t.datetime "updated_at", null: false
     t.string "role", default: "user", null: false
     t.string "remember_digest"
+    t.string "confirmation_token"
+    t.datetime "confirmed_at"
     t.index ["email"], name: "index_users_on_email", unique: true
   end