+++ spec/controllers/guild_invites_controller_spec.rb
@@ -3,6 +3,7 @@ RSpec.describe GuildInvitesController, type: :controller do
     let!(:guild) { create :guild }
 
     it_behaves_like 'User Auth'
+    it_behaves_like 'Unconfirmed User Auth'
 
     context 'for logged user' do
       sign_in_user