+++ spec/controllers/static_invites_controller_spec.rb
@@ -3,6 +3,7 @@ RSpec.describe StaticInvitesController, type: :controller do
     let!(:guild) { create :guild }
 
     it_behaves_like 'User Auth'
+    it_behaves_like 'Unconfirmed User Auth'
 
     context 'for logged user' do
       sign_in_user
@@ -76,6 +77,7 @@ RSpec.describe StaticInvitesController, type: :controller do
     let!(:guild) { create :guild }
 
     it_behaves_like 'User Auth'
+    it_behaves_like 'Unconfirmed User Auth'
 
     context 'for logged user' do
       sign_in_user