+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.8.4] - 2019-11-18
+## Unreleased
+### Added
+- mail settings with SendGrid
+- sending confirmation email
+- confirmation page
+
+## [0.8.4] - 2019-11-20
 ### Added
 - new dns name - https://guild-hall.org